requirements.txt
#requirements.txt
snowflake_snowpark-python
streamlit
requests
pandas